change: Remove support for C++ and rename files to use C suffixes
