+++ src/CMakeLists.txt
@@ -1,10 +1,10 @@
-set(COMMON_SOURCES "main.cpp" "slip.cpp")
+set(COMMON_SOURCES "main.c" "slip.c")
 
 target_sources(${TARGET_NAME} PRIVATE ${COMMON_SOURCES})
 
 if(${TARGET_CHIP} STREQUAL "esp8266")
     # zlib library will be available in esp-stub-lib, so miniz probably won't be needed.
-    add_library(miniz_obj OBJECT miniz.cpp)
+    add_library(miniz_obj OBJECT miniz.c)
 
     target_compile_options(miniz_obj PRIVATE
         # Disable specific warnings that miniz cannot satisfy
