+++ src/main.c
@@ -4,11 +4,9 @@
  * SPDX-License-Identifier: Apache-2.0 OR MIT
  */
 
+#include <stddef.h>
 #include <esp-stub-lib/flash.h>
-#include <esp-stub-lib/rom_wrappers.h>
-#include "slip.hpp"
-
-extern "C" void esp_main(void) __attribute__((used));
+#include "slip.h"
 
 #ifdef ESP8266
 __asm__(
@@ -20,10 +18,9 @@ __asm__(
     "j esp_main;");
 #endif //ESP8266
 
-extern "C" void esp_main(void)
+void esp_main(void)
 {
-    // Initialize flash subsystem
-    void *flash_state = nullptr;
+    void *flash_state = NULL;
     stub_lib_flash_init(&flash_state);
 
     // Send OHAI greeting to signal stub is active
